backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Data validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0

# Redis & Job Queue
redis==5.0.1
rq==1.16.1

# HTTP client
httpx==0.26.0

# AWS SDK
boto3==1.34.34

# CORS
python-multipart==0.0.6

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
